Variables and Data Types.py
This Python file outlines variables and data types in a python file with two challenges to complete. After practicing these concepts you should be familiar with variables, the most common Python Data Types, how to use the print statement to display the datatype of a variable as well as the value of a particular variable and how to comment out specific lines of code.